SI-6476 Accept escaped quotes in interp strings
Everyone wants to say `s"\"hello,\" world"`.

This commit considers `\"` when looking for the end of
a string part. `\\` is permitted to enable `\\"` for
literal `\"`.

However, the interpolator still receives the source text.

The wrinkly edge case is that trailing backslash must be
escaped. For instance, an interpolator that creates `File`
objects would require `F"~\Documents\\"`.

Improve error message for unterminated string to highlight
quoted quote behavior.