SI-2991 Enable untupled overload res for Java
Java will disambiguate `f(A)` and `f(A, A*)` even though
they are patently ambiguous to the Scala programmer where
A is effectively Object.

This commit adds a compiler option `-Yno-tupled-ambiguity`
to mean: if an overload seems ambiguous and one of the
alternatives is a Java method, try again with tupling disabled.

Since the library author defines the intentional disambiguation
of an overload, the new option allows the library consumer to
respect the intended disambiguation, and indeed, to enable
any disambiguation at all by Scala.