SI-2991 No Java tupling
Don't auto-tuple arguments to a Java method.

No one writes a Java method that takes an Object
or type parameter intending the user to supply a
Scala tuple.

A compiler option `-Yno-java-tupling` ensures that
tupling does not make a Java method applicable for
any purpose.

Java will disambiguate `f(A)` and `f(A, A*)` even though
they are patently ambiguous to the Scala programmer where
A is effectively Object. Turning off Java tupling will
allow invoking the single-parameter method defined in Java.

Given `f(Object)` and `f(A, B)`, Java disallows `f(x, y, z)`.
With tupling disabled, Scala also rejects that application,
which is otherwise a source of programmer error.

The compiler option can be considered for standardization
in 2.12.