SI-9123 More coherent trees with patmat, dependent types
The pattern matcher needs to substitute references to bound
variables with references to either a) synthetic temporary vals,
or to b) selections. The latter occurs under -optimize to avoid
to be frugal with local variable slots.

For instance:

```
def test(s: Some[String]) = s match {
  case Some(elem) => elem.length
}
```

Is translated to:

```
def test(s: Some[String]): Int = {
  case <synthetic> val x1: Some[String] = s;
  case4(){
    if (x1.ne(null))
      matchEnd3(x1.x.length())
    else
      case5()
  };
  case5(){
    matchEnd3(throw new MatchError(x1))
  };
  matchEnd3(x: Int){
    x
  }
}
```

However, for a long time this translation failed to consider
references to the binder in types. #4122 tried to address this
by either using standard substitution facilities where available
(references to temp vals), and by expanding the patmat's
home grown substitution to handle the more complex case of
referencing a selection.

However, this left the tree in an incoherent state; while it
patched up the `.tpe` field of `Tree`s, it failed to modify the
info of `Symbol`-s.

This led to a crash in the later uncurry phase under
`-Ydelambdafy:method`.

This commit modifies the info of such symbols to get rid of stray
refeferences to the pattern binder symbols.