Introduce sbt build
It mimics the ant build as closely as necessary to compare the produced
artifacts between the two builds, to build confidence in the transition:

 - all projects are built into the same directories in build/quick
 - include legacy projects: forkjoin, asm, actors
 - TODO: Include forkjoin & asm into library & compiler jars, respectively.

We don't carry over ant's built-in bootstrapping process;
this will be scripted outside of the build, publishing layers locally.

NOTE: the root project cannot be named `scala`:
it shadows the `scala` package in e.g., `projectConsole`.