Initial sbt definition for Scala.
It builds the library, reflect, compiler, forkjoin and asm with correct
classpaths. All projects are built into the same directories in build/quick
where Ant is building so the location of produced classes doesn't change.

Forkjoin and asm will be packaged into separate jars but we want to
distribute them the library's and compiler's jars respectively. This should
be fixed later on.

This project doesn't do a lot of other things like building docs, running
various tests, etc.