Package jars into build/pack/lib directory.
Override default location for jars to be build/pack/lib. This turned out
to be a bit more involved than I expected. Check the diff for details.

Strip down the -version suffix from jar names.

For now we're trying to mimic Ant's build as closely as possible. Ant does
not embed version numbers in jar names it creates in build/pack/lib
directory.