Performance improvement: collectFirst in TraversableOnce
collectFirst was implemented in TraversableOnce by calling toIterator and then using a non-local return to pull out a Some when the partial function succeeded.  This had two problems:

  1. If the TraversableOnce was Iterator or Iterable, stepping through until pf is happy is much (15x!) faster.
  2. If the TraversableOnce was not, creating an Iterator is a waste of time and memory

This fixes both of these issues by inspecting the self-type and choosing the appropriate implementation.

Further (modest) improvements likely possible in 2.12 by moving specialized implementations to child classes and using `applyOrElse` on the partial function with a package-private object instead of a locally created one.