Generate .properties files that are placed in jars.
Add an sbt task that generates .properties files that are packaged
into jars. The properties file contains Scala version in various
formats.

The task is registered as a resource generator so sbt can properly
track generated files and make sure they are included in packaged
jars.

The implementation of generateVersionPropertiesFile contains a note
on arbitrary order of written properties. We could consider using
https://github.com/etiennestuder/java-ordered-properties
to guarantee a stable order. Or simply ditch using Properties class
and generate a String. Our property files are simple enough that
nothing more fancy is needed.