SI-8362: Support for Future without sun.misc.Unsafe.
Instead of using AtomicReferenceFieldUpdater extends directly AtomicReference as AtomicReferenceFieldUpdater.compareAndSet() isn't necessarily even inlined by Hotspot, while AtomicReference.compareAndSet() should be compiled down to the machine's CAS instruction.
The implementation is similar to the 2.12.x branch’s one.

Add an exclusion rule for mima to not check backward and forward compatibility as the class is package protected and in an imp package.