Change project's base to src/$name
Previously, Project.base was set to ./$name and we tweaked manually the
`baseDirectory` setting to point at `.`. This was confusing sbt and its
plugins. We are setting Project.base to ./src/$name and leave
`baseDirectory` intact (so it will point at `./src/$name`.

Adapt settings that depend on `baseDirectory` accordingly.

Refactor configuration as a subproject into a common method.

Given that `baseDirectory` is pointing at `src/$name`, we don't have to
worry about `sourcesInBase`.