Change project's base to src/$name
Previously, Project.base was set to ./$name and we tweaked manually the
`baseDirectory` setting to point at `.`. This was confusing sbt and its
plugins. We are setting Project.base to ./src/$name and leave
`baseDirectory` intact (so it will point at `./src/$name`.

Settings that were dependent on `baseDirectory` has been adapted
accordingly.

Also, we refactored configuration as a subproject into a common method.