SI-7730 Fix scoping in pattern typing
Pattern binders introduced by a `CaseDef` may not be referred to by
terms or types in other patterns in the same case. The implementation
of this was too simplistic, To enforce this, `typedIdent` simply
skipped the first enclosing context when in a `PATTERNmode` or
`TYPEPATmode`.

However this was too simplistic. If the type pattern included an
existential type, the nested context containing the quantified
types was dropped, rather than that of the CaseDef. This results in
a "symbol not found" error in the enclosed pos test. The enclosed
neg tests show the related symptom: we can actually subvert the
restriction and refer to binders from within an existential
type, or from within a refined type.

The commit enforces the unusual scoping rules faithfully by:

  - introduces ContextMode.Pattern, and mark the `CaseDef`-s
    context with this while typechecking its patterns
  - within `Context.lookupSymbol`, search for the closest enclosing
    context so-marked, and ignore any definitions owned by its scope.

We can't predicate this logic on typing `Mode`-s alone. I tried
that as my first attempt, but found that in the body of a refined
type (e.g in the third part of the neg test), we have switched out
of the TYPEPATmode. Using a `ContextMode` instead lets us simply
walk the context chain and find the scope to ignore.