SI-9167 Clarify ScalaVersion parsing
Probably it was unintended to accept "2.." and "2.-11.7".

This commit makes it a bit more regular and also accepts arbitrary
text after the dash in the build string, including newlines.

Since the number parts must be numbers, accept only digits.
That also disallows "2.+11.7", which could be misconstrued as
some sort of range.

As before, the special build string prefixes "rc" and "m" are
case-insensitive, but "FINAL" is not.