Add "scala-" prefix to the name of a subproject.
Scala artifacts (like library or compiler) are published are published
with "scala-" prefix in their name. We keep project ids short so we can
refer to projects simply by e.g. `library` and not `scala-library`.

Given that `name` settings will have a different value, we switch to
`Project.id` when constructing paths for target directories.

We had to tweak properties generation because over there we use project
name as a file name without "scala-" prefix.

Since our projects are not named uniformly (e.g., "repl" and not "scala-repl"),
add a setting for each subproject that requires special handling.