SI-9134 Verify Stream#withFilter#map lazy in tail
As discussed in comments on [the issue](https://issues.scala-lang.org/browse/SI-9134),
it appears that the fix for SI-8990 also fixed this issue.

These tests demonstrate that it is fixed and will serve
as regression tests. Also, if back-ported to 2.11.x, they
could drive a potentially different fix there.

Squashed:

  - incorporate review comments from adriaanm