SI-9188 Issue a missing unchecked warning for class type param
CheckabilityChecker determined that the match was unsound, but
stumbled when trying to report which part of the type was to blame.
When this happens, the warning is debug logged, but not issued.

It failed to report that the reference to the enclosing class's
type paramter was to blame because of some code that sought to
avoid spurious warnings for dummy arguments (the kind that result
from calling `tpe_*` on a type constructor.)

I tightened up the check to re-use some existing code that does
a better job at identifying dummies. Now unchecked warnings
are issued for the last two patterns in the enclosed test.