SI-8831 Fix mapping case accessor lookup with backticked idents
Case class parameters may be declared as non-public. If this is the
case, the case accessor method is still synthesized as public, but
it has a mangled name of the form `origName$N`.

See also SI-8944 / #4125.

The compiler maps from case fields to the corresponding accessors
with this naming convention in mind. However, it is too lenient, and
assumes that only one accessor will be of the form "origName\$.*".

If, however, a regular, public case parameter is named `a b`, it will
have a name `a$u0020b`, which spurously matches that overly permissive
pattern.

This commit changes tightens up the logic to expect "origName\$\d+".