SI-8790 Detuplify match on tuple under -Xexperimental
Optimize

```
(e1,..., eN) match {
  case (p1, ..., pN) if G => B
  ...
  case _ if G => B
}
```

to

```
val t1 = e1
...
val tN = eN

t1 | ... | TN match {
  case p1 | ... | pN if G => B
  ...
  case _ if G => B
}
```

Here, `|` is truly a magic class: its constructor is a no-op,
and its accessors are local variable accesses (the `ti`).

The optimization supports scrutinees of the shape:
  - `(e1,...,eN)` (by the time we're looking at it, desugared to the next one)
  - `TupleN.apply(e1,...,eN)`
  - `new TupleN(e1,...,eN)`

Stuff like this is not allowed (can't omit the effect of the `???`)
  - `{???; TupleN}.apply(e1,...,eN)` // !isQualifierSafeToElide(fun)

TODO: this needs to be tested more thoroughly before it can
be un-Xexperimentalized (although I think I fixed the breakage in CSE
that blocked the PR last time -- see next commit).