SI-8359 Emit invokedynamic for lambdas
Suitable lambdas are identified in Delambdafy and marked with
 such with a tree annotation that includes the
data needed by the backend to emit an invokedynamic instruction.

GenBCode to rewrite instantiation of such anonymous
function classes with an invokedynamic instruction. At this
stage, I don't plan to merge the support for this into GenASM.

Between these points, the lambda capture is represented as an
application of a dummy factory symbol:

```
<dummy>(captures...) : FunctionN
```

Demo:

```
% wget http://central.maven.org/maven2/org/scala-lang/modules/scala-java8-compat_2.11/0.3.0/scala-java8-compat_2.11-0.3.0.jar
% qscala -classpath scala-java8-compat_2.11-0.3.0.jar -Ydelambdafy:method -target:jvm-1.8 -Ybackend:GenBCode
Welcome to Scala version 2.11.6-20150309-144147-c91c978c81 (Java HotSpot(TM) 64-Bit Server VM, Java 1.8.0_25).
Type in expressions to have them evaluated.
Type :help for more information.

scala> (() => "").getClass
res0: Class[_ <: () => String] = class $$Lambda$1/871790326
```

I have also corrected an error in a previous commit. The newly added
symbol test, `isDelambdafyTarget`, needs to check for the `ARTIFACT`
flag, as that is what is added to the method by `Uncurry`.