Rework subprojects name handling.
Before this commit, we used to have common way of handling names of
subproject. We would construct the name by prepending "scala-" suffix to
project id (which is simply the name of directory in src/). It turns out,
our projects are not named uniformly. For example, we have just "repl" and
not "scala-repl".

In this commit, we give up on central naming logic and have manual settings
for each subproject that requires special handling.