Add generation of shell scripts.
Shell scripts are generated with `mkBin` sbt task that calls ScalaTool.
ScalaTool is defined in project/ and is a port of Ant task that lives
in scala.tools.ant.ScalaTool. While porting, we've simplified the code
significantly.

The `mkBin` task is defined in a new subproject: dist. This subproject
will become responsible for building the distribution in the future.